feat/refactor: 포트폴리오 작성 완료 후 recoil-persist 저장 내역 삭제, 가독성 개선 - useRecoilValue 관련 코드 커스텀 훅으로 분리
